Added master password system, clear-all option, and input validation
Implemented a master password system to restrict access to the password manager. Also added a 'clear all' option for deleting stored credentials, and tweaked the input prompts to reject empty responses.

Honestly, I tried to do more than I was ready for, and the code turned out messy but it works, which honestly surprises me. I’ll rework this in a cleaner version later, but for now, I’m putting it out as-is.